requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0

spacy>=3.7.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl

presidio-analyzer>=2.2.352
presidio-anonymizer>=2.2.352

azure-ai-contentsafety>=1.0.0
azure-core>=1.30.0
python-dotenv>=1.0.1
google-generativeai>=0.8.0

regex>=2023.10.3
